+++ doc/news/changes/incompatibilities/20211019Munch
@@ -1,0 +1,4 @@
+Renamed: The namespace Functions::LevelSet has been renamed to
+Functions::SignedDistance.
+<br>
+(Peter Munch, 2021/10/19)
