+++ source/base/CMakeLists.txt
@@ -32,7 +32,7 @@ SET(_unity_include_src
   function.cc
   function_cspline.cc
   function_derivative.cc
-  function_level_set.cc
+  function_signed_distance.cc
   function_lib.cc
   function_lib_cutoff.cc
   function_parser.cc
@@ -131,7 +131,7 @@ SET(_inst
   bounding_box.inst.in
   data_out_base.inst.in
   function.inst.in
-  function_level_set.inst.in
+  function_signed_distance.inst.in
   function_restriction.inst.in
   function_time.inst.in
   function_tools.inst.in
